back/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
pillow==12.0.0
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
django-environ==0.12.0
python-dotenv==1.2.1

# pour le formatage et le linting 
black
ruff 